Archive/DataAPI/Archive/requirements.txt
BeautifulSoup==3.2.1
certifi==2017.4.17
chardet==3.0.4
fix-yahoo-finance==0.0.18
idna==2.5
matplotlib==1.1.1
multitasking==0.0.4
numpy==1.13.1
pandas==0.20.3
pandas-datareader==0.4.0
python-dateutil==2.6.1
pytz==2017.2
QSTK==0.2.6
requests==2.18.1
requests-file==1.4.2
requests-ftp==0.3.1
scikit-learn==0.13
scipy==0.11.0
simplejson==3.10.0
six==1.10.0
statsmodels==0.4.3
urllib==1.21.1
urllib2==1498656401.94
urllib3==1.21.1
yahoo-finance==1.4.0
